pyproject.toml
[project]
name = "reg-trans"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jan Grosser", email = "jan.grosser@bytepitch.com"},
]
dependencies = [
    "slack-bolt>=1.18.1",
    "python-dotenv>=1.0.1",
    "openai>=1.12.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}



[tool.pdm]
distribution = false
